avoid gcc format error on embedded builds
when using gcc 8.3.0 on Debian unstable and doing an embedded build,
I'm getting the following error:

> fsm.c:621:40: error: format '%ld' expects argument of type
> 'long int', but argument 6 has type 'time_t {aka long long int}'
> [-Werror=format=]

Let's avoid that...

+++ src/fsm.c
@@ -620,7 +620,7 @@ static int state_chg(struct osmo_fsm_inst *fi, uint32_t new_state,
 			else
 				snprintf(trailer, sizeof(trailer), "(keeping " OSMO_T_FMT
 					  ", %ld.%03lds remaining)", OSMO_T_FMT_ARGS(fi->T),
-					  remaining.tv_sec, remaining.tv_usec / 1000);
+					  (long) remaining.tv_sec, remaining.tv_usec / 1000);
 		} else if (timeout_ms) {
 			if (timeout_ms % 1000 == 0)
 				/* keep log output legacy compatible to avoid autotest failures */
